MVCCAllocator uses new MAP containers

+++ src/tools/tests/vector/vector_test_suite.hpp
@@ -26,8 +26,8 @@ public:
 
     VectorTestSuite(): TestSuite("VectorSuite")
     {
-        registerTask(new VectorTest<UByte>("UByte"));
-//        registerTask(new VectorTest<Int>("Int"));
+//        registerTask(new VectorTest<UByte>("UByte"));
+        registerTask(new VectorTest<Int>("Int"));
 //        registerTask(new VectorTest<BigInt>("BigInt"));
 
 //        registerTask(new VectorTransferTest<UByte>("Transfer"));
